requirements.txt
torch
transformers
datasets
accelerate
bitsandbytes
peft
sentence-transformers
faiss-cpu
scikit-learn
nltk
evaluate
pandas
tqdm
numpy
pyyaml
scipy